src: bump RSA/DSA parameter sizes to `size_t`
Also related internal types in openssl.c.

To reduce internal casts and match most modern crypto APIs.

Closes #2266

+++ src/hostkey.c
@@ -120,11 +120,8 @@ static int hostkey_method_ssh_rsa_init(LIBSSH2_SESSION *session,
     if(!ssh2_eob(&buf))
         return -1;
 
-    if(ssh2_rsa_new(&rsactx,
-                    e, (unsigned long)e_len,
-                    n, (unsigned long)n_len,
-                    NULL, 0, NULL, 0, NULL, 0,
-                    NULL, 0, NULL, 0, NULL, 0))
+    if(ssh2_rsa_new(&rsactx, e, e_len, n, n_len,
+                    NULL, 0, NULL, 0, NULL, 0, NULL, 0, NULL, 0, NULL, 0))
         return -1;
 
     *abstract = rsactx;
@@ -513,12 +510,7 @@ static int hostkey_method_ssh_dss_init(LIBSSH2_SESSION *session,
     if(!ssh2_eob(&buf))
         return -1;
 
-    if(ssh2_dsa_new(&dsactx,
-                    p, (unsigned long)p_len,
-                    q, (unsigned long)q_len,
-                    g, (unsigned long)g_len,
-                    y, (unsigned long)y_len,
-                    NULL, 0))
+    if(ssh2_dsa_new(&dsactx, p, p_len, q, q_len, g, g_len, y, y_len, NULL, 0))
         return -1;
 
     *abstract = dsactx;
